Change type of IResponse.SizeBytes from int to int64 (#559) (#1)
* Change type of IResponse.SizeBytes from int to int64 (#559)

This will enable counting data sizes larger than 2 GB.

* Also change type of DataTarnsferStats.MinBytes, MeanBytes, MaxBytes from int to int64 (#559)

* Also change type of DataTransferStats.Percent50,75,95,99 from int to int64 (#559)

+++ src/NBomber.Contracts/Contracts.fs
@@ -11,14 +11,14 @@ open NBomber.Contracts.Stats
 type IResponse =
     abstract StatusCode: string
     abstract IsError: bool
-    abstract SizeBytes: int
+    abstract SizeBytes: int64
     abstract LatencyMs: float
     abstract Message: string     
 
 type Response<'T> = {
     StatusCode: string
     IsError: bool
-    SizeBytes: int
+    SizeBytes: int64
     LatencyMs: float
     Message: string
     Payload: 'T option
